+++ src/bot.ts
@@ -1,11 +1,10 @@
 require("dotenv").config();
-import { BaseInteraction, Client, Events, Message } from 'discord.js';
+import { BaseInteraction, Client, ColorResolvable, Events, Guild, GuildMember, Message, Role } from 'discord.js';
 import { CommandHelper } from './utilities/CommandHelper';
 import { MongoConnection } from './services/MongoConnection';
-import { StartupUtility } from './utilities/StartupUtility';
+import { BotEventsHelper } from './services/BotEventsHelper';
 import { DiscordConnection } from './services/DiscordConnection';
 import { EmojiTracker } from './services/EmojiTracker';
-import v8 from "v8"
 
 // deploy commands anew
 CommandHelper.deployCommands();
@@ -14,17 +13,19 @@ DiscordConnection.getInstance().then((client: Client) => {
 	// When the client is ready, run this code once (https://tinyurl.com/5cavyafu)
 	client.once(Events.ClientReady, async client => {
 		console.log(`Ready! Logged in as ${client.user.tag}`);
-		console.log(v8.getHeapStatistics().heap_size_limit + " bytes");
+
+		// required for auto role update
+		BotEventsHelper.cacheUsers();
 		if(await MongoConnection.getInstance()){
 			// update servers and the database if worth doing
-			if(await StartupUtility.isNotableUpdate()){
+			if(await BotEventsHelper.isNotableUpdate()){
 				// there's no point on updating the db's data unless the major/minor has changed
-				StartupUtility.updateVersion();
-				StartupUtility.notifyServersOfBella();
+				BotEventsHelper.updateVersion();
+				BotEventsHelper.notifyServersOfBella();
 			}
 
 			// restart monitors
-			StartupUtility.restartMonitors();
+			BotEventsHelper.restartMonitors();
 		}
 	});
 
@@ -65,9 +66,15 @@ DiscordConnection.getInstance().then((client: Client) => {
 
 	// Parse messages - ONLY WORKS SINCE UNVERIFIED BOT, at 100 servers it requires verification
 	client.on(Events.MessageCreate, async (message: Message<boolean>) => {
-		if(message.author.bot) return;
-		
-		EmojiTracker.updateEmojiCountFromMessage(message);
+		if(!message.author.bot)	
+			EmojiTracker.updateEmojiCountFromMessage(message);
 	});
-})
+	
+	// Parse user updates - ONLY WORKS SINCE UNVERIFIED BOT, additionally it only works on cached users...
+	client.on(Events.UserUpdate, async (oldUser, newUser) => {
+		if(oldUser.partial)
+			oldUser = await oldUser.fetch();
+		BotEventsHelper.updateRoles(oldUser, newUser);
+	})
+});