+++ src/services/BotEventsHelper.ts
@@ -1,15 +1,34 @@
 import { version, versionToArray } from "../commands/slashCommands/version";
 
-import { MongoConnection } from "../services/MongoConnection";
-import { Client, AttachmentBuilder } from 'discord.js';
+import { MongoConnection } from "./MongoConnection";
+import { Client, AttachmentBuilder, User, ColorResolvable } from 'discord.js';
 import { PropertyDocument } from "../data/Property";
-import { ImageMaker } from "../services/ImageMaker";
-import { DiscordConnection } from "../services/DiscordConnection";
+import { ImageMaker } from "../utilities/ImageMaker";
+import { DiscordConnection } from "./DiscordConnection";
 import { MonitorDocument } from "../data/Monitor";
 import { startMonitor } from "../commands/slashCommands/monitor";
+import { getAverageColor } from "fast-average-color-node";
 
 
-export class StartupUtility {
+export class BotEventsHelper {
+    /**
+     * caches all users of guilds Bella belongs to
+     * @returns true if successful
+     */
+    public static async cacheUsers(): Promise<Boolean> {
+        try{
+            const discordClient = await DiscordConnection.getInstance();
+            discordClient.guilds.cache.forEach(async guild => {
+                await guild.members.fetch();
+            });
+        }
+        catch(error){
+            console.error(error);
+            return false;
+        }
+        return true;
+    }
+
     /**
      * Check if Bella's major/minor numbers have changed
      * @throws (Error) If mongo connection fails
@@ -89,5 +108,34 @@ export class StartupUtility {
             return false;
         }
         return true;
+    }
+
+    public static async updateRoles(oldUser: User, newUser: User){
+        if(oldUser.displayAvatarURL() === newUser.displayAvatarURL()) return;
+        const discordClient = await DiscordConnection.getInstance();
+
+		// look through guilds for this user and update their display role's colour
+		for(let guild of discordClient.guilds.cache.values()){
+			const userAsMember = (await guild.members.fetch()).find(user => newUser.id === user.id);
+			if(!userAsMember) continue;
+
+			// roles are listed top to bottom
+			// find the first with a colour which only has this member in it
+			const topRole = userAsMember.roles.cache.find(role => role.color !== 0 && role.members.size === 1);
+			if(!topRole) continue;
+
+			const color = (await getAverageColor(newUser.displayAvatarURL(), {algorithm: "simple"})).hex as ColorResolvable;
+
+			try{
+				topRole.setColor(color)
+			}
+			catch(error) {
+				const msg = `Bella could not set the colour of role ${topRole.name} in server ${guild.name} for user`;
+				console.error(msg);
+				guild.systemChannel?.send(msg).catch(error => {
+					console.error("Bella failed to warn server of role problem!");
+				})
+			};
+		}
     }
 }